[Feature #158589597]Refactor code
