+++ api/routes.py
@@ -5,8 +5,8 @@
 """
 from flask import Blueprint
 from flask_restful import Api
-from api.users.views import GetAllUsers, BorrowOps, BorrowHistory, PromoteUser
-from api.admin.views import AddBook, BookOps
+from api.users.views import GetAllUsers, BorrowOps, BorrowHistory
+from api.admin.views import AddBook, BookOps, PromoteUser
 from api.books.views import GetBooks, GetBook
 from api.auth.views import Register, Login, Logout, ResetPassword
